requirements.txt
numpy==1.26.2
opencv-python==4.8.1.78
Pillow==10.1.0
RPi.GPIO==0.7.1
spidev==3.6
ST7735==0.0.5
